Add and improve tests
